EX 10 Complete
Given a text file, write a program (console application) that counts the number of words that end with the letter t or e (not case-sensitive). We’ll say that a t or e is at the end of a word if there is not an alphabetic letter immediately following it. The program should output the resulting count.
Credit: http://codingbat.com/prob/p199171